Move Notifier interface to prebuilt-tasks module
