+++ workflow-examples/src/main/java/com/redhat/parodos/examples/move2kube/move2kubeWorkFlowConfiguration.java
@@ -12,11 +12,11 @@ import com.redhat.parodos.examples.move2kube.task.Move2KubePlan;
 import com.redhat.parodos.examples.move2kube.task.Move2KubeRetrieve;
 import com.redhat.parodos.examples.move2kube.task.Move2KubeTask;
 import com.redhat.parodos.examples.move2kube.task.Move2KubeTransform;
+import com.redhat.parodos.infrastructure.Notifier;
 import com.redhat.parodos.tasks.git.GitCloneTask;
 import com.redhat.parodos.workflow.annotation.Checker;
 import com.redhat.parodos.workflow.annotation.Infrastructure;
 import com.redhat.parodos.workflow.consts.WorkFlowConstants;
-import com.redhat.parodos.workflow.task.infrastructure.Notifier;
 import com.redhat.parodos.workflows.workflow.ParallelFlow;
 import com.redhat.parodos.workflows.workflow.SequentialFlow;
 import com.redhat.parodos.workflows.workflow.WorkFlow;
