+++ workflow-examples/src/main/java/com/redhat/parodos/examples/prebuilt/migrationtoolkit/MigrationAssessmentWorkflow.java
@@ -3,6 +3,7 @@ package com.redhat.parodos.examples.prebuilt.migrationtoolkit;
 import java.net.URI;
 import java.util.List;
 
+import com.redhat.parodos.infrastructure.Notifier;
 import com.redhat.parodos.tasks.migrationtoolkit.CreateApplicationTask;
 import com.redhat.parodos.tasks.migrationtoolkit.GetAnalysisTask;
 import com.redhat.parodos.tasks.migrationtoolkit.GetApplicationTask;
@@ -12,7 +13,6 @@ import com.redhat.parodos.workflow.annotation.Checker;
 import com.redhat.parodos.workflow.annotation.Parameter;
 import com.redhat.parodos.workflow.option.WorkFlowOption;
 import com.redhat.parodos.workflow.parameter.WorkParameterType;
-import com.redhat.parodos.workflow.task.infrastructure.Notifier;
 import com.redhat.parodos.workflows.workflow.WorkFlow;
 
 import org.springframework.beans.factory.annotation.Value;
