+++ workflow-examples/src/test/java/com/redhat/parodos/examples/move2kube/task/Move2KubeTransformTest.java
@@ -2,8 +2,8 @@ package com.redhat.parodos.examples.move2kube.task;
 
 import java.util.UUID;
 
+import com.redhat.parodos.infrastructure.Notifier;
 import com.redhat.parodos.utils.RestUtils;
-import com.redhat.parodos.workflow.task.infrastructure.Notifier;
 import com.redhat.parodos.workflow.utils.WorkContextUtils;
 import com.redhat.parodos.workflows.work.WorkContext;
 import com.redhat.parodos.workflows.work.WorkReport;
