+++ workflow-service/src/main/java/com/redhat/parodos/NotificationSender.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 import javax.ws.rs.core.HttpHeaders;
 
+import com.redhat.parodos.infrastructure.Notifier;
 import com.redhat.parodos.notification.sdk.api.ApiClient;
 import com.redhat.parodos.notification.sdk.api.ApiException;
 import com.redhat.parodos.notification.sdk.api.NotificationMessageApi;
 import com.redhat.parodos.notification.sdk.model.NotificationMessageCreateRequestDTO;
 import com.redhat.parodos.security.SecurityUtils;
-import com.redhat.parodos.workflow.task.infrastructure.Notifier;
 import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.beans.factory.annotation.Value;
